requirements.txt
requests
python-dotenv
pytest
argparse
schedule
